Rename mock logging adapter file
