+++ sts-mock/src/services/logging/tests/mockLoggingAdapter.ts
