+++ sts-mock/src/token/tokenHandler.test.ts
@@ -1,5 +1,5 @@
 import { Logger } from "../services/logging/logger";
-import { MockLoggingAdapter } from "../services/logging/tests/mockLogger";
+import { MockLoggingAdapter } from "../services/logging/tests/mockLoggingAdapter";
 import { buildLambdaContext } from "../testUtils/mockContext";
 import { registeredLogs } from "./registeredLogs";
 import { lambdaHandlerConstructor } from "./tokenHandler";
@@ -7,18 +7,20 @@ import { lambdaHandlerConstructor } from "./tokenHandler";
 describe("Token", () => {
   describe("Given lambdaHandler is called", () => {
     it("Returns 200 response", async () => {
-      const mockLogger = new MockLoggingAdapter();
+      const mockLoggingAdapter = new MockLoggingAdapter();
       const dependencies = {
         env: {},
-        logger: () => new Logger(mockLogger, registeredLogs),
+        logger: () => new Logger(mockLoggingAdapter, registeredLogs),
       };
       const result = await lambdaHandlerConstructor(
         dependencies,
         buildLambdaContext(),
       );
 
-      expect(mockLogger.getLogMessages()[0].logMessage.message).toBe("STARTED");
-      expect(mockLogger.getLogMessages()[1].logMessage.message).toBe(
+      expect(mockLoggingAdapter.getLogMessages()[0].logMessage.message).toBe(
+        "STARTED",
+      );
+      expect(mockLoggingAdapter.getLogMessages()[1].logMessage.message).toBe(
         "COMPLETED",
       );
